aiops/healing-agent/requirements.txt
# ============================================================================
# HEALING AGENT DEPENDENCIES
# ============================================================================

# Web framework
flask==3.0.0
gunicorn==21.2.0

# Kubernetes client
kubernetes==28.1.0

# Groq AI client + dependencies (FIXED VERSIONS)
groq==0.11.0
httpx==0.27.0
httpx-sse==0.4.0

# Prometheus client
prometheus-client==0.19.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Date/time
python-dateutil==2.8.2
